Python2.5 related cleanups

+++ Modules/objc/OC_PythonData.h
@@ -21,7 +21,7 @@
 @interface OC_PythonData : NSData
 {
 	PyObject* value;
-	unsigned buffer_len;
+	Py_ssize_t buffer_len;
 	const void *buffer;
 }
